+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@
  * New: Added a protected method `Register(IEventApplier)` to
    `AggregateRoot<,>` that enables developers to override how events are
    applied. Use this to e.g. implement a state object
+ * New: Allow `AggregateRoot<,>` methods for applying events, i.e.,
+  `Apply(...)`, are now allowed to be `private` and `protected`
  * New: Made `AggregateRoot<,>.Emit(...)` protected and virtual to allow
    overrides that e.g. add a standard set of metadata from the aggregate state.
  * New: Made `AggregateRoot<,>.ApplyEvent(...)` protected and virtual to