+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
  * Breaking: Updated NuGet reference `Newtonsoft.Json` to v7.0.1
    (up from v6.0.8)
+ * Breaking: Remove the empty constructor from `SingleValueObject<>`
  * New: Added a protected method `Register(IEventApplier)` to
    `AggregateRoot<,>` that enables developers to override how events are
    applied. Use this to e.g. implement a state object